Move BAT eligibility evaluation into ingest
