+++ tests/unit/bat/test_transform.py
@@ -658,261 +658,6 @@ def test_extract_group_value_returns_none_for_empty_group_value():
     assert transform.extract_group_value(soup, "Category") is None
 
 
-def test_evaluate_listing_eligibility_rejects_missing_or_unparseable_year():
-    soup = BeautifulSoup("<html></html>", "html.parser")
-
-    assert transform.evaluate_listing_eligibility(soup, "factory-five-cobra-replica") == (
-        False,
-        "listing ID year missing",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_pre_1946_listing_id():
-    soup = BeautifulSoup("<html></html>", "html.parser")
-
-    assert transform.evaluate_listing_eligibility(soup, "1941-ford-super-deluxe-coupe") == (
-        False,
-        "year before 1946",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_missing_country():
-    soup = BeautifulSoup("<html><body></body></html>", "html.parser")
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (
-        False,
-        "listing outside US",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_non_usa_country():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">CAN</span>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (
-        False,
-        "listing outside US",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_excluded_category():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/parts/">
-                    <strong class="group-title-label">Category</strong>
-                    Parts
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (
-        False,
-        "excluded category: Parts",
-    )
-
-def test_evaluate_listing_eligibility_rejects_when_any_category_is_excluded():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/convertible/">
-                    <strong class="group-title-label">Category</strong>
-                    Convertibles
-                </a>
-                <a class="group-link" href="/parts/">
-                    <strong class="group-title-label">Category</strong>
-                    Parts
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (
-        False,
-        "excluded category: Parts",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_projects_via_category():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/project/">
-                    <strong class="group-title-label">Category</strong>
-                    Projects
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911-coupe") == (
-        False,
-        "excluded category: Projects",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_race_cars_via_category():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/race-car/">
-                    <strong class="group-title-label">Category</strong>
-                    Race Cars
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1997-porsche-911-gt2-evo") == (
-        False,
-        "excluded category: Race Cars",
-    )
-
-
-def test_evaluate_listing_eligibility_rejects_replica_when_listing_id_year_missing():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/convertible/">
-                    <strong class="group-title-label">Category</strong>
-                    Convertibles
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "shelby-cobra-replica") == (
-        False,
-        "listing ID year missing",
-    )
-
-
-def test_evaluate_listing_eligibility_does_not_reject_missing_category():
-    soup = BeautifulSoup(
-        '<html><body><span class="show-country-name">USA</span></body></html>',
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (True, None)
-
-
-def test_evaluate_listing_eligibility_uses_listing_id_when_title_has_no_year():
-    soup = BeautifulSoup(
-        '<html><body><span class="show-country-name">USA</span></body></html>',
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "2010-am-general-hmmwv-military-5") == (
-        True,
-        None,
-    )
-
-def test_evaluate_listing_eligibility_does_not_reject_empty_category():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/category/">
-                    <strong class="group-title-label">Category</strong>
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (True, None)
-
-
-def test_evaluate_listing_eligibility_does_not_reject_non_excluded_category():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/convertible/">
-                    <strong class="group-title-label">Category</strong>
-                    Convertibles
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-porsche-911s-coupe") == (True, None)
-
-def test_evaluate_listing_eligibility_allows_multiple_non_excluded_categories():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/convertible/">
-                    <strong class="group-title-label">Category</strong>
-                    Convertibles
-                </a>
-                <a class="group-link" href="/truck-4x4/">
-                    <strong class="group-title-label">Category</strong>
-                    Truck & 4x4
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-ford-f-250") == (True, None)
-
-
-def test_evaluate_listing_eligibility_keeps_truck_and_4x4_in_scope():
-    soup = BeautifulSoup(
-        """
-        <html>
-            <body>
-                <span class="show-country-name">USA</span>
-                <a class="group-link" href="/truck-4x4/">
-                    <strong class="group-title-label">Category</strong>
-                    Truck & 4x4
-                </a>
-            </body>
-        </html>
-        """,
-        "html.parser",
-    )
-
-    assert transform.evaluate_listing_eligibility(soup, "1967-ford-f-250-4x4") == (True, None)
-    
 def test_get_listing_details_valid():
     html_content = """
     <html>
